refactor(bare-import-allowlist): tighten trait surface and share test harness
- Default `Rule::apply` to `Vec::new()` so lint-only rules implement only the half they emit
- Drop `BareImportAllowlist`'s empty `apply` override and unused `Edit` import
- Hoist `build_pipeline`, `fixture_config`, and `sidecar_contents` into `tests/common/mod.rs`
- Rewire `tests/diagnostics.rs` and `tests/integration.rs` onto the shared helpers
