+++ tests/common/mod.rs
@@ -3,8 +3,11 @@
 #![allow(dead_code)]
 
 use std::ffi::OsStr;
+use std::io::ErrorKind;
 use std::path::Path;
 
+use prose::config::Config;
+use prose::pipeline::Pipeline;
 use serde::Deserialize;
 use similar::TextDiff;
 
@@ -22,10 +25,27 @@ struct Sidecar {
     harness: HarnessOptions,
 }
 
+pub(crate) fn build_pipeline(directory: &str, config: &Config) -> Pipeline {
+    match directory {
+        "composition" | "suppression" => Pipeline::with_defaults(config),
+        "binding_analysis" | "identity" => Pipeline::empty(),
+        _ => Pipeline::for_rule(directory, config)
+            .unwrap_or_else(|| panic!("no rule registered for fixture directory `{directory}`")),
+    }
+}
+
 pub(crate) fn case_stem(path: &Path) -> &str {
     path.file_stem()
         .and_then(OsStr::to_str)
         .expect("fixture path has a stem")
+}
+
+pub(crate) fn fixture_config(path: &Path) -> Config {
+    sidecar_contents(path)
+        .map(|c| {
+            Config::from_pyproject_str(&c).unwrap_or_else(|e| panic!("parse sidecar config: {e}"))
+        })
+        .unwrap_or_default()
 }
 
 pub(crate) fn in_snapshot_dir(directory: &str, f: impl FnOnce()) {
@@ -42,6 +62,18 @@ pub(crate) fn parse_harness_options(contents: &str) -> HarnessOptions {
     toml::from_str::<Sidecar>(contents)
         .unwrap_or_else(|e| panic!("parse sidecar harness section: {e}"))
         .harness
+}
+
+pub(crate) fn sidecar_contents(path: &Path) -> Option<String> {
+    let stem = case_stem(path)
+        .strip_suffix(".input")
+        .expect("fixture path ends in .input.py");
+    let sidecar = path.with_file_name(format!("{stem}.config.toml"));
+    match fs_err::read_to_string(&sidecar) {
+        Ok(c) => Some(c),
+        Err(e) if e.kind() == ErrorKind::NotFound => None,
+        Err(e) => panic!("read sidecar: {e}"),
+    }
 }
 
 pub(crate) fn unified_diff(expected: &str, actual: &str) -> String {
